WIP refactor submit-review
