+++ client/app/intake/actions/ama.js
@@ -1,8 +1,50 @@
 import { ACTIONS, ENDPOINT_NAMES } from '../constants';
 import ApiUtil from '../../util/ApiUtil';
-import { formatRatingData } from '../util';
+import { formatRatingData, prepareReviewData } from '../util';
+import { formatDateStringForApi } from '../../util/DateUtil';
+import _ from 'lodash';
 
 const analytics = true;
+
+export const submitReview = (intakeId, intakeData, intakeType) => (dispatch) => {
+  dispatch({
+    type: ACTIONS.SUBMIT_REVIEW_START,
+    meta: { analytics }
+  });
+
+  const data = prepareReviewData(intakeData, intakeType);
+
+  return ApiUtil.patch(`/intake/${intakeId}/review`, { data }, ENDPOINT_NAMES.REVIEW_INTAKE).
+    then(
+      () => dispatch({
+        type: ACTIONS.SUBMIT_REVIEW_SUCCEED,
+        meta: { analytics }
+      }),
+      (error) => {
+        const responseObject = JSON.parse(error.response.text);
+        const responseErrorCodes = responseObject.error_codes;
+
+        dispatch({
+          type: ACTIONS.SUBMIT_REVIEW_FAIL,
+          payload: {
+            responseErrorCodes
+          },
+          meta: {
+            analytics: (triggerEvent, category, actionName) => {
+              triggerEvent(category, actionName, 'any-error');
+
+              _.forEach(
+                responseErrorCodes,
+                (errorVal, errorKey) => triggerEvent(category, actionName, `${errorKey}-${errorVal}`)
+              );
+            }
+          }
+        });
+
+        throw error;
+      }
+    );
+};
 
 export const completeIntake = (intakeId, intakeData) => (dispatch) => {
   dispatch({
