+++ client/app/intake/actions/appeal.js
@@ -16,47 +16,3 @@ export const setDocketType = (docketType) => ({
     }
   }
 });
-
-export const submitReview = (intakeId, appeal) => (dispatch) => {
-  dispatch({
-    type: ACTIONS.SUBMIT_REVIEW_START,
-    meta: { analytics }
-  });
-
-  const data = {
-    docket_type: appeal.docketType,
-    receipt_date: formatDateStringForApi(appeal.receiptDate),
-    claimant: appeal.claimant
-  };
-
-  return ApiUtil.patch(`/intake/${intakeId}/review`, { data }, ENDPOINT_NAMES.REVIEW_INTAKE).
-    then(
-      () => dispatch({
-        type: ACTIONS.SUBMIT_REVIEW_SUCCEED,
-        meta: { analytics }
-      }),
-      (error) => {
-        const responseObject = JSON.parse(error.response.text);
-        const responseErrorCodes = responseObject.error_codes;
-
-        dispatch({
-          type: ACTIONS.SUBMIT_REVIEW_FAIL,
-          payload: {
-            responseErrorCodes
-          },
-          meta: {
-            analytics: (triggerEvent, category, actionName) => {
-              triggerEvent(category, actionName, 'any-error');
-
-              _.forEach(
-                responseErrorCodes,
-                (errorVal, errorKey) => triggerEvent(category, actionName, `${errorKey}-${errorVal}`)
-              );
-            }
-          }
-        });
-
-        throw error;
-      }
-    );
-};
