+++ client/app/intake/actions/higherLevelReview.js
@@ -28,48 +28,3 @@ export const setSameOffice = (sameOffice) => ({
     }
   }
 });
-
-export const submitReview = (intakeId, higherLevelReview) => (dispatch) => {
-  dispatch({
-    type: ACTIONS.SUBMIT_REVIEW_START,
-    meta: { analytics }
-  });
-
-  const data = {
-    informal_conference: higherLevelReview.informalConference,
-    same_office: higherLevelReview.sameOffice,
-    receipt_date: formatDateStringForApi(higherLevelReview.receiptDate),
-    claimant: higherLevelReview.claimant
-  };
-
-  return ApiUtil.patch(`/intake/${intakeId}/review`, { data }, ENDPOINT_NAMES.REVIEW_INTAKE).
-    then(
-      () => dispatch({
-        type: ACTIONS.SUBMIT_REVIEW_SUCCEED,
-        meta: { analytics }
-      }),
-      (error) => {
-        const responseObject = JSON.parse(error.response.text);
-        const responseErrorCodes = responseObject.error_codes;
-
-        dispatch({
-          type: ACTIONS.SUBMIT_REVIEW_FAIL,
-          payload: {
-            responseErrorCodes
-          },
-          meta: {
-            analytics: (triggerEvent, category, actionName) => {
-              triggerEvent(category, actionName, 'any-error');
-
-              _.forEach(
-                responseErrorCodes,
-                (errorVal, errorKey) => triggerEvent(category, actionName, `${errorKey}-${errorVal}`)
-              );
-            }
-          }
-        });
-
-        throw error;
-      }
-    );
-};
