--- client/app/intake/actions/supplementalClaim.js
@@ -1,49 +1,0 @@
-import { ACTIONS, ENDPOINT_NAMES } from '../constants';
-import ApiUtil from '../../util/ApiUtil';
-import { formatDateStringForApi } from '../../util/DateUtil';
-import _ from 'lodash';
-
-const analytics = true;
-
-export const submitReview = (intakeId, supplementalClaim) => (dispatch) => {
-  dispatch({
-    type: ACTIONS.SUBMIT_REVIEW_START,
-    meta: { analytics }
-  });
-
-  const data = {
-    receipt_date: formatDateStringForApi(supplementalClaim.receiptDate),
-    claimant: supplementalClaim.claimant
-  };
-
-  return ApiUtil.patch(`/intake/${intakeId}/review`, { data }, ENDPOINT_NAMES.REVIEW_INTAKE).
-    then(
-      () => dispatch({
-        type: ACTIONS.SUBMIT_REVIEW_SUCCEED,
-        meta: { analytics }
-      }),
-      (error) => {
-        const responseObject = JSON.parse(error.response.text);
-        const responseErrorCodes = responseObject.error_codes;
-
-        dispatch({
-          type: ACTIONS.SUBMIT_REVIEW_FAIL,
-          payload: {
-            responseErrorCodes
-          },
-          meta: {
-            analytics: (triggerEvent, category, actionName) => {
-              triggerEvent(category, actionName, 'any-error');
-
-              _.forEach(
-                responseErrorCodes,
-                (errorVal, errorKey) => triggerEvent(category, actionName, `${errorKey}-${errorVal}`)
-              );
-            }
-          }
-        });
-
-        throw error;
-      }
-    );
-};
